WIP aqc feature flag
