+++ Cargo.toml
@@ -38,7 +38,7 @@ wildcard_imports = "warn"
 aranya-aqc-util = { version = "0.7.0", features = ["alloc"] }
 aranya-capi-core = { version = "0.4.4" }
 aranya-capi-codegen = { version = "0.3.3" }
-aranya-crypto = { version = "0.9.0", features = ["alloc", "aqc", "fs-keystore", "clone-aead", "std", "tls"] }
+aranya-crypto = { version = "0.9.0", features = ["alloc", "fs-keystore", "clone-aead", "std", "tls"] }
 aranya-crypto-ffi = { version = "0.12.0" }
 aranya-device-ffi = { version = "0.12.0" }
 aranya-envelope-ffi = { version = "0.12.0" }
