+++ crates/aranya-client-capi/Cargo.toml
@@ -21,6 +21,8 @@ doctest = false
 [features]
 default = []
 
+aqc = ["aranya-client/aqc"]
+
 
 [dependencies]
 aranya-client = { workspace = true }
