+++ crates/aranya-client-capi/src/imp/aqc.rs
@@ -1,3 +1,5 @@
+#![cfg(feature = "aqc")]
+
 use core::{
     mem::{self, MaybeUninit},
     ptr,
