+++ crates/aranya-client-capi/src/imp/config.rs
@@ -17,6 +17,7 @@ pub(crate) use team::*;
 #[derive(Clone, Debug)]
 pub struct ClientConfig {
     daemon_addr: *const c_char,
+    #[cfg(feature = "aqc")]
     aqc: AqcConfig,
 }
 
@@ -25,6 +26,7 @@ impl ClientConfig {
         self.daemon_addr
     }
 
+    #[cfg(feature = "aqc")]
     pub(crate) fn aqc_addr(&self) -> *const c_char {
         self.aqc.addr
     }
@@ -38,6 +40,7 @@ impl Typed for ClientConfig {
 #[derive(Clone, Debug)]
 pub struct ClientConfigBuilder {
     daemon_addr: *const c_char,
+    #[cfg(feature = "aqc")]
     aqc: Option<AqcConfig>,
 }
 
@@ -48,6 +51,7 @@ impl ClientConfigBuilder {
     }
 
     /// Set the config to be used for AQC
+    #[cfg(feature = "aqc")]
     pub fn aqc(&mut self, cfg: AqcConfig) {
         self.aqc = Some(cfg);
     }
@@ -69,12 +73,15 @@ impl Builder for ClientConfigBuilder {
             return Err(InvalidArg::new("daemon_addr", "field not set").into());
         }
 
-        let Some(aqc) = self.aqc else {
+        #[cfg(feature = "aqc")]
+        let Some(aqc) = self.aqc
+        else {
             return Err(InvalidArg::new("aqc", "field not set").into());
         };
 
         let cfg = ClientConfig {
             daemon_addr: self.daemon_addr,
+            #[cfg(feature = "aqc")]
             aqc,
         };
         Self::Output::init(out, cfg);
@@ -86,63 +93,81 @@ impl Default for ClientConfigBuilder {
     fn default() -> Self {
         Self {
             daemon_addr: ptr::null(),
+            #[cfg(feature = "aqc")]
             aqc: None,
         }
     }
 }
 
-/// AQC configuration.
-#[derive(Clone, Debug)]
-pub struct AqcConfig {
-    /// Address to bind AQC server to.
-    pub addr: *const c_char,
-}
+#[cfg(feature = "aqc")]
+pub use aqc::*;
+#[cfg(feature = "aqc")]
+mod aqc {
+    use std::ffi::c_char;
 
-impl Typed for AqcConfig {
-    const TYPE_ID: TypeId = TypeId::new(0x64CEB3F4);
-}
+    use aranya_capi_core::{
+        prelude::*,
+        safe::{TypeId, Typed},
+        InvalidArg,
+    };
 
-/// Builder for an [`AqcConfig`]
-#[derive(Clone, Debug)]
-pub struct AqcConfigBuilder {
-    /// Address to bind AQC server to.
-    addr: *const c_char,
-}
+    use crate::{api::defs, imp::Error};
 
-impl AqcConfigBuilder {
-    /// Sets the network address that the AQC server should
-    /// listen on.
-    pub fn addr(&mut self, addr: *const c_char) {
-        self.addr = addr;
+    /// AQC configuration.
+    #[derive(Clone, Debug)]
+    pub struct AqcConfig {
+        /// Address to bind AQC server to.
+        pub addr: *const c_char,
     }
-}
 
-impl Builder for AqcConfigBuilder {
-    type Output = defs::AqcConfig;
-    type Error = Error;
+    impl Typed for AqcConfig {
+        const TYPE_ID: TypeId = TypeId::new(0x64CEB3F4);
+    }
 
-    /// # Safety
-    ///
-    /// No special considerations.
-    unsafe fn build(self, out: &mut MaybeUninit<Self::Output>) -> Result<(), Self::Error> {
-        if self.addr.is_null() {
-            return Err(InvalidArg::new("addr", "field not set").into());
+    /// Builder for an [`AqcConfig`]
+    #[derive(Clone, Debug)]
+    pub struct AqcConfigBuilder {
+        /// Address to bind AQC server to.
+        addr: *const c_char,
+    }
+
+    impl AqcConfigBuilder {
+        /// Sets the network address that the AQC server should
+        /// listen on.
+        pub fn addr(&mut self, addr: *const c_char) {
+            self.addr = addr;
         }
-
-        let cfg = AqcConfig { addr: self.addr };
-
-        Self::Output::init(out, cfg);
-        Ok(())
     }
-}
 
-impl Typed for AqcConfigBuilder {
-    const TYPE_ID: TypeId = TypeId::new(0x153AE387);
-}
+    impl Builder for AqcConfigBuilder {
+        type Output = defs::AqcConfig;
+        type Error = Error;
 
-impl Default for AqcConfigBuilder {
-    fn default() -> Self {
-        Self { addr: ptr::null() }
+        /// # Safety
+        ///
+        /// No special considerations.
+        unsafe fn build(self, out: &mut MaybeUninit<Self::Output>) -> Result<(), Self::Error> {
+            if self.addr.is_null() {
+                return Err(InvalidArg::new("addr", "field not set").into());
+            }
+
+            let cfg = AqcConfig { addr: self.addr };
+
+            Self::Output::init(out, cfg);
+            Ok(())
+        }
+    }
+
+    impl Typed for AqcConfigBuilder {
+        const TYPE_ID: TypeId = TypeId::new(0x153AE387);
+    }
+
+    impl Default for AqcConfigBuilder {
+        fn default() -> Self {
+            Self {
+                addr: core::ptr::null(),
+            }
+        }
     }
 }
