+++ crates/aranya-client-capi/src/imp/mod.rs
@@ -3,6 +3,7 @@ pub mod client;
 pub mod config;
 pub mod error;
 
+#[cfg(feature = "aqc")]
 pub use aqc::*;
 pub use client::*;
 pub use config::*;
