+++ crates/aranya-client/Cargo.toml
@@ -14,6 +14,10 @@ workspace = true
 [features]
 default = []
 
+aqc = [
+    "aranya-daemon-api/aqc"
+]
+
 # Enable debugging.
 debug = []
