+++ crates/aranya-daemon-api/Cargo.toml
@@ -12,6 +12,8 @@ rust-version.workspace = true
 [lints]
 workspace = true
 
+[features]
+aqc = ["aranya-crypto/aqc"]
 
 [dependencies]
 aranya-crypto = { workspace = true }
