+++ crates/aranya-daemon/src/daemon.rs
@@ -8,11 +8,14 @@ use aranya_crypto::{
     Engine, Rng,
 };
 use aranya_keygen::{KeyBundle, PublicKeys};
+#[cfg(feature = "aqc")]
+use aranya_runtime::StorageProvider;
 use aranya_runtime::{
     storage::linear::{libc::FileManager, LinearStorageProvider},
-    ClientState, StorageProvider,
+    ClientState,
 };
 use aranya_util::ready;
+#[cfg(feature = "aqc")]
 use bimap::BiBTreeMap;
 use buggy::{bug, Bug, BugExt};
 use ciborium as cbor;
@@ -20,10 +23,10 @@ use serde::{de::DeserializeOwned, Serialize};
 use tokio::{fs, sync::Mutex, task::JoinSet};
 use tracing::{error, info, info_span, Instrument as _};
 
+#[cfg(feature = "aqc")]
+use crate::{actions::Actions, aqc::Aqc};
 use crate::{
-    actions::Actions,
-    api::{ApiKey, DaemonApiServer, EffectReceiver, QSData},
-    aqc::Aqc,
+    api::{ApiKey, DaemonApiServer, DaemonApiServerArgs, EffectReceiver, QSData},
     aranya,
     config::{Config, Toggle},
     keystore::{AranyaStore, LocalStore},
@@ -186,16 +189,17 @@ impl Daemon {
             .await?;
             let local_addr = sync_server.local_addr()?;
 
-            let graph_ids = client
-                .aranya
-                .lock()
-                .await
-                .provider()
-                .list_graph_ids()?
-                .flatten()
-                .collect::<Vec<_>>();
-
+            #[cfg(feature = "aqc")]
             let aqc = {
+                let graph_ids = client
+                    .aranya
+                    .lock()
+                    .await
+                    .provider()
+                    .list_graph_ids()?
+                    .flatten()
+                    .collect::<Vec<_>>();
+
                 let peers = {
                     let mut peers = BTreeMap::new();
                     for graph_id in &graph_ids {
@@ -227,20 +231,21 @@ impl Daemon {
                 aranya_store,
             };
 
-            let api = DaemonApiServer::new(
+            let api = DaemonApiServer::new(DaemonApiServerArgs {
                 client,
                 local_addr,
-                cfg.uds_api_sock(),
-                api_sk,
-                pks,
+                uds_path: cfg.uds_api_sock(),
+                sk: api_sk,
+                pk: pks,
                 peers,
                 recv_effects,
-                invalid_graphs,
+                invalid: invalid_graphs,
+                #[cfg(feature = "aqc")]
                 aqc,
                 crypto,
                 seed_id_dir,
-                Some(data),
-            )?;
+                quic: Some(data),
+            })?;
             Ok(Self {
                 sync_server,
                 syncer,
