+++ examples/rust/Cargo.toml
@@ -38,7 +38,7 @@ unwrap_used = "warn"
 wildcard_imports = "warn"
 
 [dependencies]
-aranya-client = { path = "../../crates/aranya-client" }
+aranya-client = { path = "../../crates/aranya-client", features = ["aqc"] }
 aranya-daemon-api = { path = "../../crates/aranya-daemon-api" }
 aranya-util = { path = "../../crates/aranya-util" }
