+++ examples/rust/run.bash
@@ -22,7 +22,8 @@ echo "Building aranya-daemon..."
 cargo build \
     --release \
     --manifest-path Cargo.toml \
-    --bin aranya-daemon
+    --bin aranya-daemon \
+    --features aqc
 
 daemon="${current_dir}/target/release/aranya-daemon"
 example="${current_dir}/examples/rust/target/release/aranya-example"
